move relayServerMessage to utils
